ban to send, edit, delete message for locked, email sending if user locked or unlocked, front

+++ src/main/java/com/example/socialNetwork/controller/MessageController.java
@@ -4,7 +4,9 @@ import com.example.socialNetwork.domain.Message;
 import com.example.socialNetwork.domain.User;
 import com.example.socialNetwork.domain.Views;
 import com.example.socialNetwork.dto.MessagePageDto;
+import com.example.socialNetwork.repo.UserRepo;
 import com.example.socialNetwork.service.MessageService;
+import com.example.socialNetwork.service.UserService;
 import com.fasterxml.jackson.annotation.JsonView;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -13,6 +15,7 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import java.io.IOException;
+import java.time.LocalDateTime;
 
 @RestController
 @RequestMapping("message")
@@ -20,10 +23,12 @@ public class MessageController {
     public static final int MESSAGES_PER_PAGE = 11  ;
 
     private final MessageService messageService;
+    private final UserService userService;
 
     @Autowired
-    public MessageController(MessageService messageService) {
+    public MessageController(MessageService messageService, UserService userService) {
         this.messageService = messageService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -45,6 +50,18 @@ public class MessageController {
     public Message create(@RequestBody Message message,
                           @AuthenticationPrincipal User user
     ) throws IOException {
+        User userDb = userService.findById(user.getId());
+
+        if (!userDb.isAccountNonLocked()){
+            Message errMessage = new Message();
+            User errUser = userService.findById("bot");
+
+            errMessage.setCreationDate(LocalDateTime.now());
+            errMessage.setText(user.getName() + ", Ваш аккаунт был заблокирован из-за нарушения правил нашего чата, вы можете написать администратору, либо связаться с нами по почте");
+
+            return messageService.create(errMessage, errUser);
+        }
+
         return messageService.create(message, user);
     }
 
@@ -52,13 +69,39 @@ public class MessageController {
     @JsonView(Views.FullMessage.class)
     public Message update(
             @PathVariable("id") Message messageFromDb,
-            @RequestBody Message message
+            @RequestBody Message message,
+            @AuthenticationPrincipal User user
     ) throws IOException {
+        User userDb = userService.findById(user.getId());
+
+        if (!userDb.isAccountNonLocked()){
+            Message errMessage = new Message();
+            User errUser = userService.findById("bot");
+
+            errMessage.setCreationDate(LocalDateTime.now());
+            errMessage.setText(user.getName() + ", Ваш аккаунт был заблокирован из-за нарушения правил нашего чата, вы можете написать администратору, либо связаться с нами по почте");
+
+            return messageService.create(errMessage, errUser);
+        }
+
         return messageService.update(messageFromDb, message);
     }
 
     @DeleteMapping("{id}")
-    public void  delete(@PathVariable("id") Message message){
+    public void  delete(@PathVariable("id") Message message, @AuthenticationPrincipal User user) throws IOException {
+        User userDb = userService.findById(user.getId());
+
+        if (!userDb.isAccountNonLocked()){
+            Message errMessage = new Message();
+            User errUser = userService.findById("bot");
+
+            errMessage.setCreationDate(LocalDateTime.now());
+            errMessage.setText(user.getName() + ", Ваш аккаунт был заблокирован из-за нарушения правил нашего чата, вы можете написать администратору, либо связаться с нами по почте");
+
+            messageService.create(errMessage, errUser);
+            return;
+        }
+
         messageService.delete(message);
     }
